+++ scorex-transaction/src/main/scala/scorex/transaction/TransactionSettings.scala
@@ -1,6 +1,7 @@
 package scorex.transaction
 
 import play.api.libs.json.JsObject
+import scala.concurrent.duration._
 
 trait TransactionSettings {
   val settingsJSON: JsObject
@@ -11,4 +12,7 @@ trait TransactionSettings {
   private val DefaultMaxRollback = 100
   lazy val MaxRollback = (settingsJSON \ "maxRollback").asOpt[Int].getOrElse(DefaultMaxRollback)
 
+  private val DefaultUtxRebroadcastInterval: FiniteDuration = 30.seconds
+  lazy val utxRebroadcastInterval: FiniteDuration = (settingsJSON \ "utxRebroadcastInterval").asOpt[Int]
+    .map(x => x.seconds).getOrElse(DefaultUtxRebroadcastInterval)
 }